(vendor panel) small fix
add to view "approved"
security edit product

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -93,9 +93,9 @@ class VendorProductImageGalleryController extends Controller
         $productImage = ProductImageGallery::findOrFail($id);
 
         /** Check product vendor */
-        // if($productImage->product->vendor_id !== Auth::user()->vendor->id){
-        //     abort(404);
-        // }
+        if($productImage->product->vendor_id !== Auth::user()->vendor->id){
+            abort(404);
+        }
 
         $this->deleteImage($productImage->image);
         $productImage->delete();
